refactor(native): move context compaction into its own module
The compaction predicates (`turns_character_count`, `value_character_count`,
`should_compact`, `compaction_notice_block`), the automatic path
(`AutomaticCompactionOutcome`, `maybe_compact_accounted`,
`run_automatic_compaction`), the optimizer path and its records, the context
quality assessment, and the pre-optimizer compatibility fallback move to
`api_process_adapter/compaction.rs`.

Measured at 858 lines, not the 808 design.md estimated: the estimate's range
started after the four predicates above, which tasks.md 3.4 assigns here.
